Revert "SW-4960 Lazy load routes in terraware web instead of prefetching before we know the user principal context (#2259)"
This reverts commit f1a1ec1961a5fcdae26d597b5b10c2feffb47fbb.
